+++ src/main/scala/com/github/oen9/slinky/bridge/reactkonva/Line.scala
@@ -3,7 +3,13 @@ package com.github.oen9.slinky.bridge.reactkonva
 import com.github.oen9.slinky.bridge.konva.Konva
 import com.github.oen9.slinky.bridge.reactkonva.ReactKonvaDOM.Context
 import Konva.Vector2d
+import org.scalajs.dom.raw.DragEvent
+import org.scalajs.dom.raw.Event
 import org.scalajs.dom.raw.HTMLImageElement
+import org.scalajs.dom.raw.MouseEvent
+import org.scalajs.dom.raw.PointerEvent
+import org.scalajs.dom.raw.TouchEvent
+import org.scalajs.dom.raw.WheelEvent
 import scala.scalajs.js.|
 import scalajs.js
 import slinky.core.annotations.react
@@ -91,9 +97,29 @@ import slinky.core.ExternalComponentWithRefType
 
     _useStrictMode: js.UndefOr[Boolean] = js.undefined,
     draggable: js.UndefOr[Boolean] = js.undefined,
-    onClick: js.UndefOr[() => Unit] = js.undefined,
-    onDragEnd: js.UndefOr[() => Unit] = js.undefined,
-    onTap: js.UndefOr[() => Unit] = js.undefined,
+
+    onMouseOver: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onMouseMove: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onMouseOut: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onMouseEnter: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onMouseLeave: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onMouseDown: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onMouseUp: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onWheel: js.UndefOr[Konva.KonvaEventObject[WheelEvent] => Unit] = js.undefined,
+    onClick: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onDblClick: js.UndefOr[Konva.KonvaEventObject[MouseEvent] => Unit] = js.undefined,
+    onTouchStart: js.UndefOr[Konva.KonvaEventObject[TouchEvent] => Unit] = js.undefined,
+    onTouchMove: js.UndefOr[Konva.KonvaEventObject[TouchEvent] => Unit] = js.undefined,
+    onTouchEnd: js.UndefOr[Konva.KonvaEventObject[TouchEvent] => Unit] = js.undefined,
+    onTap: js.UndefOr[Konva.KonvaEventObject[Event] => Unit] = js.undefined,
+    onDblTap: js.UndefOr[Konva.KonvaEventObject[Event] => Unit] = js.undefined,
+    onDragStart: js.UndefOr[Konva.KonvaEventObject[DragEvent] => Unit] = js.undefined,
+    onDragMove: js.UndefOr[Konva.KonvaEventObject[DragEvent] => Unit] = js.undefined,
+    onDragEnd: js.UndefOr[Konva.KonvaEventObject[DragEvent] => Unit] = js.undefined,
+    onTransform: js.UndefOr[Konva.KonvaEventObject[Event] => Unit] = js.undefined,
+    onTransformStart: js.UndefOr[Konva.KonvaEventObject[Event] => Unit] = js.undefined,
+    onTransformEnd: js.UndefOr[Konva.KonvaEventObject[Event] => Unit] = js.undefined,
+    onContextMenu: js.UndefOr[Konva.KonvaEventObject[PointerEvent] => Unit] = js.undefined,
   )
   override val component = ReactKonvaDOM.Line
 }